// src/app/shared/enum/svg-urls.enum.ts
export enum NutriscoreLinks {
  NUTRISCORE_A = 'https://static.openfoodfacts.org/images/attributes/dist/nutriscore-a.svg',
  NUTRISCORE_B = 'https://static.openfoodfacts.org/images/attributes/dist/nutriscore-b.svg',
  NUTRISCORE_C = 'https://static.openfoodfacts.org/images/attributes/dist/nutriscore-c.svg',
  NUTRISCORE_D = 'https://static.openfoodfacts.org/images/attributes/dist/nutriscore-d.svg',
  NUTRISCORE_E = 'https://static.openfoodfacts.org/images/attributes/dist/nutriscore-e.svg',
  NUTRISCORE_UNKNOWN = 'https://static.openfoodfacts.org/images/attributes/dist/nutriscore-unknown.svg',
}

export enum EcoscoreLinks {
  ECOSCORE_A = 'https://static.openfoodfacts.org/images/attributes/dist/ecoscore-a.svg',
  ECOSCORE_B = 'https://static.openfoodfacts.org/images/attributes/dist/ecoscore-b.svg',
  ECOSCORE_C = 'https://static.openfoodfacts.org/images/attributes/dist/ecoscore-c.svg',
  ECOSCORE_D = 'https://static.openfoodfacts.org/images/attributes/dist/ecoscore-d.svg',
  ECOSCORE_E = 'https://static.openfoodfacts.org/images/attributes/dist/ecoscore-e.svg',
  ECOSCORE_UNKNOWN = 'https://static.openfoodfacts.org/images/attributes/dist/ecoscore-unknown.svg',
}

export enum NovagroupLinks {
  NOVAGROUP_1 = 'https://static.openfoodfacts.org/images/attributes/dist/nova-group-1.svg',
  NOVAGROUP_2 = 'https://static.openfoodfacts.org/images/attributes/dist/nova-group-2.svg',
  NOVAGROUP_3 = 'https://static.openfoodfacts.org/images/attributes/dist/nova-group-3.svg',
  NOVAGROUP_4 = 'https://static.openfoodfacts.org/images/attributes/dist/nova-group-4.svg',
  NOVAGROUP_UNKNOWN = 'https://static.openfoodfacts.org/images/attributes/dist/nova-group-unknown.svg',
}